barcode_tracker/requirements.txt
Flask
waitress
pillow
